+++ src/clj_kondo/impl/types.clj
@@ -46,6 +46,8 @@
     :transducer
     :list
     :seq
+    :lazy-seq
+    :ipending
     :sorted-map
     :boolean
     :atom
@@ -94,6 +96,7 @@
    :transducer #{:ifn :fn}
    :list #{:seq :sequential :seqable :coll :stack}
    :seq #{:seqable :sequential :coll}
+   :lazy-seq #{:seq :seqable :sequential :coll}
    :sequential #{:coll :seqable}
    :sorted-map #{:map :seqable :associative :coll :ifn :ilookup}
    :atom #{:ideref}
@@ -114,18 +117,21 @@
    :double #{:float :number}
    :number #{:neg-int :pos-int :nat-int :int :long :short :double :byte :ratio :float}
    :coll #{:map :sorted-map :vector :set :sorted-set :list :associative :seq
-           :sequential :ifn :stack :ilookup}
+           :sequential :ifn :stack :ilookup :lazy-seq}
    :seqable #{:coll :vector :set :sorted-set :map :associative
               :char-sequence :string :nil
-              :list :seq :sequential :ifn :stack :sorted-map :ilookup :array}
+              :list :seq :sequential :ifn :stack :sorted-map :ilookup :array
+              :lazy-seq}
    :associative #{:map :vector :sequential :stack :sorted-map}
    :ifn #{:fn :transducer :symbol :keyword :map :set :sorted-set :vector
           :associative :seqable :coll :sequential :stack :sorted-map :var
           :ideref :ilookup}
    :fn #{:transducer}
-   :seq #{:list :stack}
-   :stack #{:list :vector :seq :sequential :seqable :coll :ifn :associative :ilookup}
-   :sequential #{:seq :list :vector :ifn :associative :stack :ilookup}
+   :seq #{:list :stack :lazy-seq}
+   :lazy-seq #{:stack :ipending}
+   :stack #{:list :vector :seq :sequential :seqable :coll :ifn :associative
+            :ilookup :lazy-seq}
+   :sequential #{:seq :list :vector :ifn :associative :stack :ilookup :lazy-seq}
    :map #{:sorted-map}
    :set #{:sorted-set}
    :ideref #{:atom :var :ifn}
@@ -160,6 +166,8 @@
    :ratio "ratio"
    :seqable "seqable collection"
    :seq "seq"
+   :lazy-seq "lazy seq"
+   :ipending "pending (unrealized lazy seq, delay, future, or promise)"
    :vector "vector"
    :stack "stack (list, vector, etc.)"
    :associative "associative collection"
@@ -240,7 +248,6 @@
     (float) :float
     (Double java.lang.Double) :nilable/double
     (Float java.lang.Float) :nilable/float
-    (ratio?) :ratio
     (CharSequence java.lang.CharSequence) :nilable/char-sequence
     (String java.lang.String) :nilable/string ;; as this is now way to
     ;; express non-nilable,