+++ CHANGELOG.md
@@ -17,6 +17,10 @@ For a list of breaking changes, check [here](#breaking-changes).
 <!-- - [ ] update lein-clj-kondo -->
 <!-- - [ ] update carve -->
 
+## Unreleased
+
+- Add new type `:lazy-seq` and `:ipending` for better lazy sequence type checking. Lazy sequence functions like `range`, `map`, `filter`, etc. now return `:lazy-seq` type, and `realized?` has type spec that accepts `:ipending` types. ([@jramosg](https://github.com/jramosg))
+
 ## 2026.01.19
 
 - [#2735](https://github.com/clj-kondo/clj-kondo/issues/2735): NEW linter: `:duplicate-refer` which warns on duplicate entries in `:refer` of `:require`. ([@jramosg](https://github.com/jramosg))